XCF-1694 Allow for passing ProfileLocation to generateProfiles function: it's needed to generate profiles directly to WebResponse output streams.
git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@10161 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXconfig/neoconf/src/org/sipfoundry/sipxconfig/device/FileSystemProfileLocation.java
@@ -17,6 +17,7 @@ import java.io.OutputStream;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
+import org.apache.commons.lang.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
@@ -40,6 +41,9 @@ public class FileSystemProfileLocation implements ProfileLocation {
     }
 
     public void removeProfile(String profileName) {
+        if (StringUtils.isEmpty(profileName)) {
+            return;
+        }
         try {
             File file = getProfileFile(profileName);
             FileUtils.forceDelete(file);
